chore: use zustand for pedal data
